+++ config/urls.py
@@ -3,6 +3,7 @@ from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
 from django.contrib.auth import views as auth_views
+from accounts import views as accounts_views
 
 from core.views import DashboardView
 from accounts.auth import TwoFactorLoginView
@@ -10,7 +11,7 @@ from accounts.auth import TwoFactorLoginView
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("accounts/login/", TwoFactorLoginView.as_view(), name="login"),
-    path("accounts/logout/", auth_views.LogoutView.as_view(), name="logout"),
+    path("accounts/logout/", accounts_views.SignOutView.as_view(), name="logout"),
     path("accounts/password_change/", auth_views.PasswordChangeView.as_view(
         template_name="registration/password_change.html",
         success_url="/accounts/login/"), name="password_change"),